back/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
kafka-python==2.0.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiokafka==0.8.0
pgvector==0.2.0
sqlalchemy-utils==0.41.1